requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
alembic==1.13.2
email-validator==2.2.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
 